backend/requirements.txt
reflex==0.4.0
qdrant-client==1.7.5
transformers==4.37.0
torch==2.1.0
pydantic==2.5.0
rank-bm25==0.2.2
fastapi==0.109.0
uvicorn[standard]==0.27.0
neo4j==5.16.0
redis==5.0.0
sentence-transformers==2.3.0
langchain==0.1.0
langchain-community==0.0.0
openai==1.10.0
anthropic==0.18.0
PyPDF2==3.0.0
python-docx==1.1.0
python-multipart==0.0.6
Pillow==10.2.0
librosa==0.10.0
pydub==0.25.1
opencv-python==4.9.0
pydantic-settings==2.1.0
httpx==0.26.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
aiofiles==23.2.1
cohere==4.0.0
sqlalchemy==2.0.0
alembic==1.13.0
